+++ app/api/[idUser]/user/[idUserDelete]/route.ts
@@ -1,4 +1,6 @@
+import { verifyAccessToken } from "@/lib/jwt";
 import prisma from "@/lib/prisma";
+import { cookies } from "next/headers";
 import { NextResponse } from "next/server";
 
 export async function DELETE(
@@ -13,6 +15,13 @@ export async function DELETE(
         { error: "Missing idUser parameter", idUser, idUserDelete },
         { status: 400 }
       );
+    }
+
+    const cookieStore = await cookies();
+    const accessToken = cookieStore.get("accessToken")?.value || "";
+    const decoded = await verifyAccessToken(accessToken);
+    if (!decoded || decoded.id !== idUser || decoded.role !== "SELLER") {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
     }
 
     await prisma.user.delete({
